Move ts_ast nodes down a module
