+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/ts_ast/azle_type_ref/get_name.rs
