+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/ts_ast/mod.rs
@@ -1,0 +1,3 @@
+mod azle_type_ref;
+
+pub use azle_type_ref::AzleTypeRef;
